Debug and enhance update.sh Add the bin/defroute script (to add/del the default network route) Add the bin/run script (to be put in epiconcept_build@files.epiconcept.fr:.ssh) Correct exit-code in savedist.sh and do not exit on diffs Enhance send.sh (mostly add/del default route with bin/defroute) Add update.log/ to .gitignore

+++ send.sh
@@ -14,26 +14,50 @@ usr=epiconcept_build
 srv=files.epiconcept.fr
 
 test -d $dist || { echo "$Prg: cannot find '$(realpath $dist)'" >&2; exit 1; }
+test "$(id -un)" = 'php' || { echo "$Prg: must run as 'php' user" >&2; exit 1; }
 
 cleanup()
 {
-    echo "Deleting $tmp/"
+    #global Del
+    echo "Removing $tmp/"
     rm -rf $tmp
+    test "$Del" && sudo bin/defroute del && echo "Deleted default route"
 }
 
+trap cleanup 0
+
+#   Setup default route
+bin/defroute >/dev/null || {
+    sudo -l | grep $PWD/bin/defroute >/dev/null || {
+	echo "$Prg: 'sudo bin/defroute' is not configured" >&2
+	exit 2;
+    }
+    sudo bin/defroute add
+    echo "Added default route"
+    Del=y	# Delete route at exit
+}
+
+#   Copy package files
 rm -fr $tmp
 mkdir $tmp
+echo "Preparing package files"
 for deb in "$dist"/*/*/*.deb; do
     cp -p "$deb" $tmp/
 done
-trap cleanup 0
 
+#   Sync files on $srv
+#	-r	recursive
+#	-l	copy symlinks
+#	-t	preserve mtimes
+#	-u	update (skip files newer on $srv)
+#	-v	verbose
 rsync -rltuv $tmp/ $usr@$srv:/space/applisdata/php/ || {
     xc=$?
     echo "$Prg: rsync failed (xc=$xc)" >&2
     exit $xc
 }
 
+#   -t:	Force pseudo-terminal alloc
 ssh -t $usr@$srv /usr/local/bin/apt_deploy.sh || {
     xc=$?
     echo "$Prg: apt_deploy.sh on '$srv' failed (xc=$xc)" >&2;
